// frontend/src/shared/types/domain.ts
export interface VariableDefinition {
  name: string;
  label: string;
  description?: string;
  type: "string" | "number" | "boolean" | "date";
  required: boolean;
  default_value?: string | number | boolean;
}